Chore: Implement 2-digit Seven Segment Driver

+++ Electric Water Heater/Utilities/Datatype.h
@@ -11,6 +11,6 @@
 
 
 typedef unsigned char uint8_t;
-typedef unsigned short int uint16_t;
+typedef unsigned int uint16_t;
 
 #endif /* DATATYPE_H_ */
